corrected types for user_tokens table.

+++ db/src/models.rs
@@ -1,11 +1,11 @@
 use chrono::{DateTime, Utc};
-use diesel::{Identifiable, Queryable, Associations};
+use diesel::{Identifiable, Queryable, Associations, Insertable};
 use super::schema::*;
 
-#[derive(Identifiable, Queryable, Debug)]
+#[derive(Identifiable, Insertable, Queryable, Debug)]
 #[diesel(table_name = users)]
 pub struct User {
-    pub id: i32,
+    pub id: i64,
     pub email: String,
     pub hashed_password: String,
     pub confirmed_at: Option<DateTime<Utc>>,
@@ -21,11 +21,11 @@ pub enum TokenTypeEnum {
     Session
 }
 
-#[derive(Identifiable, Queryable, Associations, Debug)]
+#[derive(Identifiable, Insertable, Queryable, Associations, Debug)]
 #[diesel(table_name = user_tokens, belongs_to(User))]
 pub struct UserToken {
-    pub id: i32,
-    pub user_id: i32,
+    pub id: i64,
+    pub user_id: i64,
     pub token: Vec<u8>,
     pub token_type: TokenTypeEnum,
     pub sent_to: Option<String>,
